scripts/pipecat-examples-webrtc-docker/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Pipecat example - voice-agent"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "fastapi[all]",
    "uvicorn",
    "pipecat-ai[google,silero, webrtc]>=0.0.102",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]